+++ docs/src/modules/components/examples/CoreFeaturedExamples.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
-import FeaturedExamples from 'docs-toolpad/src/modules/components/examples/FeaturedExamples';
+import ExamplesFeatured from 'docs-toolpad/src/modules/components/examples/ExamplesFeatured';
 import coreExamples from 'docs-toolpad/src/modules/components/examples/coreExamples';
 
 export default function CoreFeaturedExamples() {
-  return <FeaturedExamples examplesFile={coreExamples()} />;
+  return <ExamplesFeatured examples={coreExamples} />;
 }
