+++ docs/src/modules/components/examples/ExamplesGrid.tsx
@@ -12,11 +12,10 @@ import IconButton from '@mui/material/IconButton';
 import Tooltip from '@mui/material/Tooltip';
 import Stack from '@mui/material/Stack';
 import SvgIcon from '@mui/material/SvgIcon';
-import type { Example } from './types';
+import { Example, versionGitHubLink } from './examplesUtils';
 
 interface ExamplesGridProps {
-  examplesFile: Example[];
-  reverse?: boolean;
+  examples: Example[];
 }
 
 function StackBlitzIcon() {
@@ -35,12 +34,10 @@ function CodeSandboxIcon() {
   );
 }
 
-function ExamplesGrid(props: ExamplesGridProps) {
+export default function ExamplesGrid(props: ExamplesGridProps) {
   const t = useTranslate();
 
-  const examples = (props.reverse ? [...props.examplesFile].reverse() : props.examplesFile).filter(
-    (example: Example) => example.featured !== true,
-  );
+  const examples = props.examples.filter((example: Example) => example.featured !== true);
   const docsTheme = useTheme();
 
   return (
@@ -73,7 +70,7 @@ function ExamplesGrid(props: ExamplesGridProps) {
                 component="a"
                 image={computedSrc}
                 title={example.description}
-                href={example.href || example.source}
+                href={versionGitHubLink(example.href || example.source)}
                 rel="nofollow"
                 sx={(theme) => ({
                   height: 0,
@@ -105,7 +102,12 @@ function ExamplesGrid(props: ExamplesGridProps) {
                   justifyContent: 'space-between',
                 }}
               >
-                <Button component="a" href={example.source} size="small" target="_blank">
+                <Button
+                  component="a"
+                  href={versionGitHubLink(example.source)}
+                  size="small"
+                  target="_blank"
+                >
                   {t('source')}
                 </Button>
                 <Stack direction="row" spacing={1}>
@@ -144,5 +146,3 @@ function ExamplesGrid(props: ExamplesGridProps) {
     </Grid>
   );
 }
-
-export default ExamplesGrid;
