--- docs/src/modules/components/examples/types.ts
@@ -1,12 +1,0 @@
-export interface Example {
-  title: string;
-  description: string;
-  src: string;
-  srcDark?: string;
-  href: string;
-  source: string;
-  codeSandbox?: boolean;
-  stackBlitz?: boolean;
-  new?: boolean;
-  featured?: boolean;
-}
